% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentimsr.R
\name{explore.tdf.tables}
\alias{explore.tdf.tables}
\title{Explore the contentents of the sqlite .tdf database.}
\usage{
explore.tdf.tables(opentims, ...)
}
\arguments{
\item{opentims}{Instance of OpenTIMS}

\item{...}{Parameters passed to head and tail functions.}
}
\value{
No return value, called for its side effect of printing the tables.
}
\description{
Explore the contentents of the sqlite .tdf database.
}
\examples{
path.d = system.file("extdata", "test.d", package = "opentimsr")
setup_opensource()
D = OpenTIMS(path.d)
if (interactive()) explore.tdf.tables(D)
CloseTIMS(D)
}
