% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentimsr.R
\name{length,OpenTIMS-method}
\alias{length,OpenTIMS-method}
\title{Get the overall number of peaks.}
\usage{
\S4method{length}{OpenTIMS}(x)
}
\arguments{
\item{x}{OpenTIMS data instance.}
}
\value{
The total number of peaks in the dataset.
}
\description{
Get the overall number of peaks.
}
\examples{
path.d = system.file("extdata", "test.d", package = "opentimsr")
setup_opensource()
D = OpenTIMS(path.d)
length(D) # number of peaks in the dataset
CloseTIMS(D)
}
