% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentimsr.R
\name{MS1}
\alias{MS1}
\title{Get MS1 frame numbers.}
\usage{
MS1(opentims)
}
\arguments{
\item{opentims}{Instance of OpenTIMS}
}
\value{
Numbers of frames corresponding to MS1, i.e. precursor ions.
}
\description{
Get MS1 frame numbers.
}
\examples{
path.d = system.file("extdata", "test.d", package = "opentimsr")
setup_opensource()
D = OpenTIMS(path.d)
MS1(D)
CloseTIMS(D)
}
