% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentimsr.R
\name{OpenTIMS}
\alias{OpenTIMS}
\title{Get OpenTIMS data representation.}
\usage{
OpenTIMS(path.d)
}
\arguments{
\item{path.d}{Path to the TimsTOF '*.d' folder containing the data (requires the folder to contain only 'analysis.tdf' and 'analysis.tdf_bin').}
}
\value{
An object of class \code{OpenTIMS}, used as the data handle by the other functions.
}
\description{
Get OpenTIMS data representation.
}
\examples{
path.d = system.file("extdata", "test.d", package = "opentimsr")
setup_opensource()
D = OpenTIMS(path.d)
D[1] # first frame
CloseTIMS(D)
}
