% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentimsr.R
\name{opentims_set_threads}
\alias{opentims_set_threads}
\title{Set the number of threads to be used for data processing by OpenTIMS}
\usage{
opentims_set_threads(n)
}
\arguments{
\item{n}{The number of worker threads to be used.}
}
\value{
No return value, called for its side effect of setting the number of threads.
}
\description{
Frames are decoded in parallel by \code{\link{query}}, \code{\link{query_slice}},
\code{\link{rt_query}}, \code{[} and \code{range}. By default all detected cores
are used; a value of 0 restores that default. When the environment variable
\code{_R_CHECK_LIMIT_CORES_} is set (as during \code{R CMD check --as-cran}),
the default is two threads.
}
\examples{
opentims_set_threads(1)
}
