% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentimsr.R
\name{peaks_per_frame_cnts}
\alias{peaks_per_frame_cnts}
\title{Get the number of peaks per frame.}
\usage{
peaks_per_frame_cnts(opentims)
}
\arguments{
\item{opentims}{Instance of OpenTIMS.}
}
\value{
Number of peaks in each frame.
}
\description{
Get the number of peaks per frame.
}
\examples{
path.d = system.file("extdata", "test.d", package = "opentimsr")
setup_opensource()
D = OpenTIMS(path.d)
peaks_per_frame_cnts(D)
CloseTIMS(D)
}
