% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentimsr.R
\name{query_slice}
\alias{query_slice}
\title{Query for raw data.}
\usage{
query_slice(opentims, from = NULL, to = NULL, by = 1, columns = all_columns)
}
\arguments{
\item{opentims}{Instance of OpenTIMS.}

\item{from}{First frame to extract.}

\item{to}{Last frame to extract.}

\item{by}{Every by-th frame gets extracted (starting from the first one).}

\item{columns}{Vector of columns to extract. Defaults to all columns.}
}
\value{
data.frame with selected columns.
}
\description{
Get the raw data from Bruker's 'tdf_bin' format.
Defaults to both raw data ('frame','scan','tof','intensity') and its tranformations into physical units ('mz','inv_ion_mobility','retention_time').
}
\details{
We assume 'from' <= 'to'.
}
\examples{
path.d = system.file("extdata", "test.d", package = "opentimsr")
setup_opensource()
D = OpenTIMS(path.d)
query_slice(D, 1, 2)    # frames 1 to 2
query_slice(D, 1, 2, 2) # every second frame between 1 and 2
query_slice(D, 1, 2, columns=c('scan','intensity')) # only 'scan' and 'intensity'
CloseTIMS(D)
}
