% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentimsr.R
\name{range,OpenTIMS-method}
\alias{range,OpenTIMS-method}
\title{Select a range of frames to extract.}
\usage{
\S4method{range}{OpenTIMS}(x, from, to, by = 1L, na.rm = FALSE)
}
\arguments{
\item{x}{OpenTIMS data instance.}

\item{from}{The first frame to extract.}

\item{to}{The last+1 frame to extract. Frame with that number will not get extracted, but some below that number might.}

\item{by}{Extract each by-th frame.}

\item{na.rm}{Needed to comply with the generic. There should not be any NA frames, 
since the from and to are checked against the limits of the data.}
}
\value{
For the selected frames, a data.frame with columns 'frame', 'scan', 'tof' and 'intensity', one row per peak.
}
\description{
This is similar to using the from:to:by operator in Python.
}
\examples{
path.d = system.file("extdata", "test.d", package = "opentimsr")
setup_opensource()
D = OpenTIMS(path.d)
head(range(D, 1, 3))    # frames 1 and 2
head(range(D, 1, 3, 2)) # every second frame, i.e. only frame 1
CloseTIMS(D)
}
