% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentimsr.R
\name{rts}
\alias{rts}
\title{Get the retention time for each frame.}
\usage{
rts(opentims)
}
\arguments{
\item{opentims}{Instance of OpenTIMS.}
}
\value{
Retention times corresponding to each frame.
}
\description{
Get the retention time for each frame.
}
