% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentimsr.R
\name{[,OpenTIMS-method}
\alias{[,OpenTIMS-method}
\title{Get some frames of data.}
\usage{
\S4method{[}{OpenTIMS}(x, i)
}
\arguments{
\item{x}{OpenTIMS data instance.}

\item{i}{Frame numbers to extract.}
}
\value{
For the requested frames, a data.frame with columns 'frame', 'scan', 'tof' and 'intensity', one row per peak.
}
\description{
Get some frames of data.
}
\examples{
path.d = system.file("extdata", "test.d", package = "opentimsr")
setup_opensource()
D = OpenTIMS(path.d)
head(D[1])   # first frame
head(D[1:2]) # frames 1 and 2
CloseTIMS(D)
}
