% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentimsr.R
\name{tables_names}
\alias{tables_names}
\title{Extract tables from sqlite database analysis.tdf.}
\usage{
tables_names(opentims)
}
\arguments{
\item{opentims}{Instance of OpenTIMS}
}
\value{
Names of tables.
}
\description{
Extract tables from sqlite database analysis.tdf.
}
\examples{
path.d = system.file("extdata", "test.d", package = "opentimsr")
setup_opensource()
D = OpenTIMS(path.d)
tables_names(D)
CloseTIMS(D)
}
