% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentimsr.R
\name{CloseTIMS}
\alias{CloseTIMS}
\title{Close the TIMS data handle and release all resources.}
\usage{
CloseTIMS(opentims)
}
\arguments{
\item{opentims}{Instance of OpenTIMS.}
}
\value{
void
}
\description{
Calling this method is not mandatory - the resources will anyway
be cleanly released when the garbage collector decides to
destroy the OpenTIMS data handle. The only purpose of this
method is to enable them to be explicitly released earlier than that.
}
\examples{
\dontrun{
D = OpenTIMS('path/to/your/folder.d')
CloseTIMS(D)
}
}
