% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentimsr.R
\name{download_bruker_proprietary_code}
\alias{download_bruker_proprietary_code}
\title{Get Bruker's code needed for running proprietary time of flight to mass over charge and scan to drift time conversion.}
\usage{
download_bruker_proprietary_code(
  target.folder,
  net_url = paste0("https://raw.githubusercontent.com/MatteoLacki/",
    "opentims_bruker_bridge/main/opentims_bruker_bridge/"),
  mode = "wb",
  ...
)
}
\arguments{
\item{target.folder}{Folder where to store the 'dll' or 'so' file.}

\item{net_url}{The url with location of all files.}

\item{mode}{Which mode to use when downloading a file?}

\item{...}{Other parameters to 'download.file'.}
}
\value{
Path to the output 'timsdata.dll' on Windows and 'libtimsdata.so' on Linux.
}
\description{
Bruker's 'timsdata' library is proprietary software owned by Bruker. It is not
part of opentimsr and is not distributed with it: this function downloads it
only when called, from "https://github.com/MatteoLacki/opentims_bruker_bridge",
and by using it you agree to the license terms stated there.
The library is available only for Linux and Windows. opentimsr works without it,
using its built-in open-source converters (see \code{\link{setup_opensource}}).
}
\examples{
\dontrun{
download_bruker_proprietary_code("your/prefered/destination/folder")
}
}
