% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentimsr.R
\name{download_bruker_proprietary_code}
\alias{download_bruker_proprietary_code}
\title{Get Bruker's code needed for running proprietary time of flight to mass over charge and scan to drift time conversion.}
\usage{
download_bruker_proprietary_code(
  target.folder,
  net_url = paste0("https://github.com/MatteoLacki/opentims_bruker_bridge/",
    "raw/main/opentims_bruker_bridge/"),
  mode = "wb",
  ...
)
}
\arguments{
\item{target.folder}{Folder where to store the 'dll' or 'so' file.}

\item{net_url}{The url with location of all files.}

\item{mode}{Which mode to use when downloading a file?}

\item{...}{Other parameters to 'download.file'.}
}
\value{
Path to the output 'timsdata.dll' on Windows and 'libtimsdata.so' on Linux.
}
\description{
By using this function you aggree to terms of license precised in "https://github.com/MatteoLacki/opentims_bruker_bridge".
The conversion, due to independent code-base restrictions, are possible only on Linux and Windows operating systems.
Works on full open-source solution are on the way.
}
