% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentimsr.R
\name{explore.tdf.tables}
\alias{explore.tdf.tables}
\title{Explore the contentents of the sqlite .tdf database.}
\usage{
explore.tdf.tables(opentims, ...)
}
\arguments{
\item{opentims}{Instance of OpenTIMS}

\item{...}{Parameters passed to head and tail functions.}
}
\description{
Explore the contentents of the sqlite .tdf database.
}
\examples{
\dontrun{
D = OpenTIMS('path/to/your/folder.d')
explore.tdf.tables(D) 
}
}
