% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentimsr.R
\name{length,OpenTIMS-method}
\alias{length,OpenTIMS-method}
\title{Get the overall number of peaks.}
\usage{
\S4method{length}{OpenTIMS}(x)
}
\arguments{
\item{x}{OpenTIMS data instance.}
}
\description{
Get the overall number of peaks.
}
