% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentimsr.R
\name{min_max_measurements}
\alias{min_max_measurements}
\title{Get border values for measurements.}
\usage{
min_max_measurements(opentims)
}
\arguments{
\item{opentims}{Instance of OpenTIMS.}
}
\value{
data.frame Limits of individual extracted quantities.
}
\description{
Get the min-max values of the measured variables (except for TOFs, that would require iteration through data rather than parsing metadata).
}
\examples{
\dontrun{
D = OpenTIMS('path/to/your/folder.d')
min_max_measurements(D) # this gives a small data-frame with min and max values.
}
}
