% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentimsr.R
\docType{class}
\name{OpenTIMS-class}
\alias{OpenTIMS-class}
\title{TimsTOF data accessor.}
\description{
S4 class that facilitates data queries for TimsTOF data.
}
\section{Slots}{

\describe{
\item{\code{path.d}}{Path to raw data folder (typically *.d).}

\item{\code{handle}}{Pointer to raw data.}

\item{\code{min_frame}}{The index of the minimal frame.}

\item{\code{max_frame}}{The index of the miximal frame.}

\item{\code{frames}}{A data.frame with information on the frames (contents of the Frames table in the sqlite db).}

\item{\code{all_columns}}{Names of available columns.}
}}

