% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentimsr.R
\name{OpenTIMS}
\alias{OpenTIMS}
\title{Get OpenTIMS data representation.}
\usage{
OpenTIMS(path.d)
}
\arguments{
\item{path.d}{Path to the TimsTOF '*.d' folder containing the data (requires the folder to contain only 'analysis.tdf' and 'analysis.tdf_bin').}
}
\description{
Get OpenTIMS data representation.
}
\examples{
\dontrun{
D = OpenTIMS(path_to_.d_folder)
D[1] # First frame.
}
}
