% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentimsr.R
\name{opentims_set_threads}
\alias{opentims_set_threads}
\title{Set the number of threads to be used for data processing by OpenTIMS}
\usage{
opentims_set_threads(n)
}
\arguments{
\item{n}{The number of worker threads to be used.}
}
\value{
void
}
\description{
A value of 0 is acceptable: it will cause OpenTIMS to use all detected cores.
}
\examples{
\dontrun{
opentims_set_threads(1)
}
}
