% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentimsr.R
\name{query_slice}
\alias{query_slice}
\title{Query for raw data.}
\usage{
query_slice(opentims, from = NULL, to = NULL, step = 1L, columns = all_columns)
}
\arguments{
\item{opentims}{Instance of OpenTIMS.}

\item{from}{First frame to extract.}

\item{to}{Last frame to extract.}

\item{step}{Every step-th frame gets extracted (starting from the first one).}

\item{columns}{Vector of columns to extract. Defaults to all columns.}
}
\value{
data.frame with selected columns.
}
\description{
Get the raw data from Bruker's 'tdf_bin' format.
Defaults to both raw data ('frame','scan','tof','intensity') and its tranformations into physical units ('mz','dt','rt').
}
\details{
We assume 'from' <= 'to'.
}
