% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentimsr.R
\name{setup_bruker_so}
\alias{setup_bruker_so}
\title{Dynamically link Bruker's DLL to enable tof-mz and scan-inv_ion_mobility conversion.}
\usage{
setup_bruker_so(path)
}
\arguments{
\item{path}{Path to the 'libtimsdata.so' on Linux or 'timsdata.dll' on Windows, as produced by 'download_bruker_proprietary_code'.}
}
\description{
By using this function you aggree to terms of license precised in "https://github.com/MatteoLacki/opentims_bruker_bridge".
The conversion, due to independent code-base restrictions, are possible only on Linux and Windows operating systems.
Works on full open-source solution are on the way.
}
\examples{
\dontrun{
so_path = download_bruker_proprietary_code("your/prefered/destination/folder")
setup_bruker_so(so_path)
}
}
