% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentimsr.R
\name{[,OpenTIMS-method}
\alias{[,OpenTIMS-method}
\title{Get some frames of data.}
\usage{
\S4method{[}{OpenTIMS}(x, i)
}
\arguments{
\item{x}{OpenTIMS data instance.}

\item{i}{An array of nonzero indices to extract.}
}
\description{
Get some frames of data.
}
\examples{
\dontrun{
D = OpenTIMS('path/to/your/folder.d')
print(head(D[10]))
print(head(D[10:100]))
}
}
