% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentimsr.R
\name{table2df}
\alias{table2df}
\title{Extract tables from sqlite database analysis.tdf.}
\usage{
table2df(opentims, names)
}
\arguments{
\item{opentims}{Instance of OpenTIMS}

\item{names}{Names to extract from the sqlite database.}
}
\value{
A list of tables.
}
\description{
Export a table from sqlite.
}
